Move createKeystone to lib
