+++ packages-next/keystone/src/index.ts
@@ -1,1 +1,1 @@
-export { createKeystone } from './classes/Keystone';
+export { createKeystone } from './lib/createKeystone';
